refactor nuernberger_land_de
- rename source: only top-level domain is necessary
- fix issue with non-ascii characters in ICS UID field
- split waste types
